requirements.txt
fastapi
uvicorn
python-jose
passlib==1.7.4
bcrypt==3.2.2
transformers
torch
sentencepiece
protobuf
streamlit
requests
